requirements.txt
requests
python-dotenv
pandas
google-cloud-bigquery
nltk
vaderSentiment
pytz
python-dateutil
yfinance
pandas-gbq